+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Support for VIFE `E001 1101` (Standard conform data content)
 - Test for LAN-WMBUS-G2-LDS/LDP
+- Support for expansion of Compact Profile with Register Numbers (Orthogonal VIFE 0x1E)
 
 ## v0.2.0